# Platfrom-UI-01/src/app/services/models/poste-response.ts
/* tslint:disable */
/* eslint-disable */
export interface PosteResponse {
  formation?: string;
  id?: number;
  mission?: string;
  nbreRecrutement?: number;
  offreId?: number;
  technologies?: string;
}