# Platfrom-UI-01/src/openapi/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://192.168.239.1:8081/api/v1",
      "description": "Generated server url"
    }
  ],
  "paths": {
    "/postes/{id}": {
      "get": {
        "tags": [
          "poste-controller"
        ],
        "operationId": "getPosteById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PosteResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "poste-controller"
        ],
        "operationId": "updatePoste",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PosteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PosteResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "poste-controller"
        ],
        "operationId": "deletePoste",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/postes": {
      "post": {
        "tags": [
          "poste-controller"
        ],
        "operationId": "createPoste",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PosteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PosteResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/postes/offre/{offreId}": {
      "get": {
        "tags": [
          "poste-controller"
        ],
        "operationId": "getPostesByOffre",
        "parameters": [
          {
            "name": "offreId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PosteResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "PosteRequest": {
        "required": [
          "formation",
          "mission",
          "technologies"
        ],
        "type": "object",
        "properties": {
          "technologies": {
            "type": "string"
          },
          "formation": {
            "type": "string"
          },
          "mission": {
            "type": "string"
          },
          "nbreRecrutement": {
            "type": "integer",
            "format": "int32"
          },
          "offreId": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "PosteResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "technologies": {
            "type": "string"
          },
          "formation": {
            "type": "string"
          },
          "mission": {
            "type": "string"
          },
          "nbreRecrutement": {
            "type": "integer",
            "format": "int32"
          },
          "offreId": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
    }
  }
}